Add action for updating a product's variations based on valid attribute combinations.
This adds a new action called `updateProductVariations` (WOOCOMMERCE_UPDATE_PRODUCT_VARIATIONS). It takes a product object, an array of existing variations for that product, and a list of which variations should now be considered valid. This allows us to update the list of variations for a product as a user enters and creates new variation types (attributes).

If an existing variation is no longer present in our list of allowed variations, it is dropped. If a variation does not exist yet, but is in our list of allowed variations, a new variation is created. Otherwise, variations (and their associated data) are preserved.